Revert "feat: add an examples page (#64)"
This reverts commit 278afefa52542110bf33df0c3293f7106149fc86.
