+++ cspell.json
@@ -3,12 +3,7 @@
   "enabled": true,
   "language": "en-US",
   "allowCompoundWords": true,
-  "dictionaries": [
-    "typescript",
-    "node",
-    "npm",
-    "html"
-  ],
+  "dictionaries": ["typescript", "node", "npm", "html"],
   "enabledLanguageIds": [
     "typescript",
     "typescriptreact",
@@ -29,6 +24,7 @@
     "horiz",
     "illu",
     "incentivizes",
+    "JWT",
     "JWK",
     "JWKs",
     "JWT",
@@ -37,7 +33,6 @@
     "merkle",
     "Mesha",
     "nullifer",
-    "Orbz",
     "petorbz",
     "POAP",
     "roboto",
