--- src/Examples/Card/scrollbar-styles.module.css
@@ -1,9 +1,0 @@
-.tags {
-  -ms-overflow-style: none;
-  scrollbar-width: none;
-  overflow-y: scroll;
-}
-
-.tags::-webkit-scrollbar {
-  display: none;
-}
