+++ src/Layout/Header/ThemeSelector/index.tsx
@@ -1,20 +1,103 @@
-import { Fragment, memo, useContext } from 'react'
+import {
+  ComponentType,
+  Fragment,
+  memo,
+  NamedExoticComponent,
+  useCallback,
+  useEffect,
+  useState,
+} from 'react'
 import { Listbox } from '@headlessui/react'
 import cn from 'classnames'
+import { Icon } from 'common/Icon'
 import { styles } from 'common/helpers/styles'
-import { DarkIcon, LightIcon, ThemeContext } from 'common/contexts/ThemeContext'
 
-export const ThemeSelector = memo(function ThemeSelector(props: {
-  className?: string
-}) {
-  const { themes, preferedTheme, changePreferedTheme } =
-    useContext(ThemeContext)
+const LightIcon = memo(function LightIcon(props: { className?: string }) {
+  return <Icon name="theme-light" className={cn('h-4 w-4', props.className)} />
+})
+const DarkIcon = memo(function LightIcon(props: { className?: string }) {
+  return <Icon name="theme-dark" className={cn('h-4 w-4', props.className)} />
+})
+const SystemIcon = memo(function LightIcon(props: { className?: string }) {
+  return <Icon name="theme-system" className={cn('h-4 w-4', props.className)} />
+})
+
+type SystemSchemes = 'light' | 'dark'
+type Schemes = SystemSchemes | 'system'
+
+const themes = [
+  {
+    name: 'Light',
+    value: 'light',
+    icon: LightIcon,
+  },
+  {
+    name: 'Dark',
+    value: 'dark',
+    icon: DarkIcon,
+  },
+  {
+    name: 'System',
+    value: 'system',
+    icon: SystemIcon,
+  },
+]
+
+export const ThemeSelector = memo(function ThemeSelector(props: { className?: string }) {
+  const [systemScheme, setSystemScheme] = useState<SystemSchemes>()
+  const [preferredScheme, setPreferredScheme] = useState<Schemes>('system')
+
+  // NOTE: listen system scheme
+  useEffect(() => {
+    const mediaQuery = window.matchMedia('(prefers-color-scheme: dark)')
+    setSystemScheme(mediaQuery.matches ? 'dark' : 'light')
+
+    const handleChangeSystemScheme = (event: MediaQueryListEvent) => {
+      setSystemScheme(event.matches ? 'dark' : 'light')
+    }
+
+    mediaQuery.addEventListener('change', handleChangeSystemScheme)
+
+    return () => {
+      mediaQuery.removeEventListener('change', handleChangeSystemScheme)
+    }
+  }, [])
+
+  // NOTE: listen change from another tab
+  useEffect(() => {
+    setPreferredScheme((localStorage.getItem('theme') ?? 'system') as Schemes)
+
+    const handleChangeStorage = (event: StorageEvent) => {
+      if (event.key !== 'theme') {
+        return
+      }
+
+      setPreferredScheme(event.newValue as Schemes)
+    }
+
+    addEventListener('storage', handleChangeStorage)
+
+    return () => {
+      removeEventListener('storage', handleChangeStorage)
+    }
+  }, [])
+
+  // NOTE: set theme to body
+  useEffect(() => {
+    document.documentElement.dataset['theme'] =
+      preferredScheme === 'system' ? systemScheme : preferredScheme
+  }, [systemScheme, preferredScheme])
+
+  const handleChangePreferredScheme = useCallback((value: Schemes) => {
+    setPreferredScheme(value)
+    localStorage.setItem('theme', value)
+  }, [])
 
   return (
     <Listbox
       as="div"
-      value={preferedTheme}
-      onChange={changePreferedTheme}
+      value={preferredScheme}
+      onChange={handleChangePreferredScheme}
       {...props}
     >
       <Listbox.Label className="sr-only">Theme</Listbox.Label>
@@ -25,7 +108,7 @@ export const ThemeSelector = memo(function ThemeSelector(props: {
       <Listbox.Options
         className={cn(
           'absolute top-full left-1/2 mt-3 w-36 -translate-x-1/2 space-y-1 rounded-lg border p-3 text-sm font-medium',
-          'border-262f41/10 bg-[#e9ebee] text-94a2b8 dark:border-262f41 dark:bg-1a2436'
+          'bg-[#e9ebee] border-262f41/10 text-94a2b8 dark:border-262f41 dark:bg-1a2436'
         )}
       >
         {themes.map((theme) => (
