--- src/pages/examples.tsx
@@ -1,2 +1,0 @@
-import { Examples } from 'Examples'
-export default Examples
